clean up code organization
